% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add.R
\name{add}
\alias{add}
\title{Add together}
\usage{
add(x, y)
}
\arguments{
\item{x}{a number.}

\item{y}{a number.}
}
\value{
A number, the value of "x" pluss the value of "y"
}
\description{
This function adds two numbers together.
}
\examples{
add(134, 288)

add(pi, 56)

}
\seealso{
\link{Arithmetic}, \code{\link{sum}}.
}
\author{
Elisabeth Knudsen, Msc. student at the University of Iceland
}
