% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arearect.R
\name{arearect}
\alias{arearect}
\title{Rectangle Area}
\usage{
arearect(x, y)
}
\arguments{
\item{x}{a number.}

\item{y}{a number.}
}
\value{
A number which is the surface area of a rectangle of with sides of length x and y.
}
\description{
Calculate the surface area of a rectangle.
}
\examples{
arearect(4,5)
 
}
\seealso{
\link{Arithmetic}, \code{\link{twice}}.
}
\author{
Clara Jégousse
}
