% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bestAIC.R
\name{bestAIC}
\alias{bestAIC}
\title{Multiply by Two}
\usage{
bestAIC(m1, m2)
}
\arguments{
\item{lm1}{and lm2 are models.}
}
\value{
A number, two times the value of x.
}
\description{
This function computes and compates de AIC value between two models.
}
\examples{
bestAIC(model1, model2)

}
\author{
Laura Aixalà Perelló
}
