% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cel2kel.R
\name{cel2kel}
\alias{cel2kel}
\title{Celsius to Kelvin}
\usage{
cel2kel(x)
}
\arguments{
\item{x}{Temperature in Celsius degrees.}
}
\description{
Converts temperatures from Celsius degrees to Kelvin scale.
}
\examples{
cel2kel(70)
cel2kel(-4)
}
\seealso{
\link{Arithmetic}
}
\author{
Vigdís Freyja Helmutsdóttir, MSc. student at the University of Iceland.
}
