% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/half.R
\name{half}
\alias{half}
\title{Divide by Two}
\usage{
half(x)
}
\arguments{
\item{x}{a number.}
}
\value{
A number, half the value of x.
}
\description{
This function divides x by 2.
}
\examples{
half(5)
 
half(pi)

 
}
\seealso{
\link{Arithmetic}, \code{\link{twice}}.
}
\author{
Ragnhildur Gudmundsdottir
}
