% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lm_plot.R
\name{lm_plot}
\alias{lm_plot}
\title{LM Plot}
\usage{
lm_plot(m)
}
\arguments{
\item{m}{linear model}
}
\value{
Scatterplot and lm regression line as well as the equation
}
\description{
Plot scatterplot with linear regression line.
}
\examples{
model <- lm(dist~speed, data=cars)
plot.lm(model)

}
\author{
Clara Jégousse
}
