% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.lm.R
\name{plot.lm}
\alias{plot.lm}
\title{Plot scatterplot with linear regression line}
\usage{
\method{plot}{lm}(m)
}
\arguments{
\item{m}{linear model}
}
\value{
Scatterplot and lm regression line as well as the equation
}
\description{
Plot scatterplot with linear regression line
}
\examples{
	
model <- lm(dist~speed, data=cars)
plot.lm(model)
}
