% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poppar.R
\name{poppar}
\alias{poppar}
\title{Population parameters from lifetable}
\usage{
poppar(x, y)
}
\arguments{
\item{x}{A 3 column table containing going from left or right age, abundance at each life stage whether relative or absolute and offspring produced at lifestage..}
}
\value{
R0,Tc,r,birth-ex, current population growth, average generation time, intrinsic population growth and life expectancy at birth (if age 0 is at the top of a given table and life stages are seperated by a constant interval, e.g. 1 year) all expressed in the given time unit..
}
\description{
This function gives various point estimates of current population growth.
}
\examples{
poppar(cbind((0:5),sort(runif(6,0,10),decreasing=TRUE),sort(c(0,0,0,runif(3,0,12)))))

poppar(cbind((0:5),sort(runif(6,0,10),decreasing=TRUE),sort(c(0,0,0,runif(3,0,12)))),20)

}
\seealso{
\href{https://cran.r-project.org/web/packages/MortalityTables/index.html}{MortailityTables}
 \href{https://cran.r-project.org/web/packages/LifeTables/index.html}{LifeTables}
}
\author{
Ölvir Styrmisson
}
