\name{poweroftwo}
\alias{poweroftwo}

\title{
The power of two
}

\usage{
poweroftwo(x)
}

\arguments{
  \item{x}{a number}
}

\value{a number, x in the power of two}

\description{This function puts x in the power of two}
}

\seealso{
\link{Aritmetric}
}

\examples{ 
poweroftwo(10)

poweroftwo(pi)
}

\author{Hulda Margrét Birkisdóttir
}

