% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r2.R
\name{r2}
\alias{r2}
\title{Calculate R2}
\usage{
r2(x, y)
}
\arguments{
\item{x}{a number.}

\item{y}{a numeric vector.}
}
\value{
A number
}
\description{
This function calculates R2 from a linear regression
}
\examples{
r2(10,5)

}
\seealso{
\link{Arithmetic}
}
\author{
Hafrún Gunnarsdóttir
}
