% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/squared.R
\name{squared}
\alias{squared}
\title{Squared}
\usage{
squared(x)
}
\arguments{
\item{x}{a number.}
}
\value{
A number, x squared.
}
\description{
This function makes x squared.
}
\examples{
squared(10)

squared(pi)

}
\seealso{
\link{Arithmetic}, \code{\link{sqrt}}.
}
\author{
Hulda Margrét Birkisdóttir
}
