% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subtract.R
\name{subtract}
\alias{subtract}
\title{Subtract y from x}
\usage{
subtract(x, y)
}
\arguments{
\item{x}{a number.}

\item{y}{a number.}
}
\value{
A number, value of x subtracted by the value of y.
}
\description{
This function subtracts y from x.
}
\examples{
subtract(10,8)

subtract(pi, (pi+1))

}
\seealso{
\link{Arithmetic}, \code{\link{twice}}
}
\author{
Sebastien Matlosz
}
