\name{twice}
\alias{twice}
\title{Multiply by Two}
\description{This function multiplies x by 2.}
\usage{
twice(x)
}
\arguments{
  \item{x}{a number.}
}
\value{a number, two times the value of x.}
\author{Arni Magnusson}

\seealso{
\link{Arithmetic}
}
\examples{
twice(10)

twice(pi)
}
