% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unlucky_twice.R
\name{unlucky_twice}
\alias{unlucky_twice}
\title{Unlucky Twice}
\usage{
unlucky_twice(x)
}
\arguments{
\item{x}{is a number}
}
\value{
\code{x} multiplied by 2 (if you're lucky!)
}
\description{
The \code{unlucky_twice} function enables you to multiply by two but in 50\% of the cases the output will be \code{"you are unlucky"}.
}
\details{
This function will only work for numerical arguments. It neither brings luck nor alters it in any way. Either you are lucky or you aren't!
}
\examples{
unlucky_twice(4)
unlucky_twice(-3)
unlucky_twice((14+9*8)+3)
unlucky_twice(pi)

}
\seealso{
\link{Arithmetic},
\code{\link{if}}
}
\author{
Jóhannes Bjarki Urbancic Tómasson
}
