% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bestAIC.R
\name{bestAIC}
\alias{bestAIC}
\title{Compare AIC}
\usage{
bestAIC(m1, m2)
}
\arguments{
\item{m1}{is a fitted model.}

\item{m2}{is another fitted model.}
}
\value{
Shows a message on the screen, indicating which model has the lower AIC.
}
\description{
This function computes and compares the AIC value between two models.
}
\examples{
model.wt <- lm(mpg~wt, data=mtcars)
model.wt.hp <- lm(mpg~wt+hp, data=mtcars)
bestAIC(model.wt, model.wt.hp)

}
\author{
Laura Aixalà Perelló
}
