% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poppar.R
\name{poppar}
\alias{poppar}
\title{Population parameters from life table}
\usage{
poppar(pop, K)
}
\arguments{
\item{pop}{A 3 column table containing age, abundance at each life stage, whether relative or absolute, and offspring produced at life stage.}

\item{K}{Estimated carrying capacity of the environment where the represented population lives.}
}
\value{
A vector containing R0, Tc, r, and birth-ex: current population growth, average generation time, intrinsic population growth and life expectancy at birth (if age 0 is at the top of a given table and life stages are separated by a constant interval, e.g. 1 year) all expressed in the given time unit.
}
\description{
This function gives various point estimates of current population growth.
}
\details{
K, carrying capacity, is optional. If it is not provided it is assumed the observed growth is not constrained by the population density.
}
\note{
Calculation of estimated parameters:

\eqn{R0 = \sum(lx * bx)}{R0 = Σ(lx * bx)}

\eqn{Gc = \sum(lx * bx * x) / R0}{Gc = Σ(lx * bx * x) / R0}

\eqn{r(1) =  ln(R0) / Gc}

\eqn{r(2) = ((dN / dt) * K) / (N * (K - N))}

\eqn{ex = Tx / lx}

Where R0 is the basic reproduction rate, lx proportion of living individuals at age x compared to age 0, bx numbers of offspring produced by individual at age x, Gc the average generation length, x the age in a given time unit, r the instantaneous multiplication rate, defined without or with K, Tx is a sum of proportion of individuals surviving at the midpoint of present life stage (here age 0) and future life stages and ex is the expected life length at age 0.

Some other needed values:

\eqn{lx = Ni / N0}

\eqn{bx = Noi / Ni}

\eqn{Lx = (lxi + lxi + 1) / 2}

\eqn{Tx = \sum Lx}{Tx = Σ Lx}

Ni is the number of individuals at a particular life stage, Noi is the offspring produced at a particular life stage and Lx is the proportion of individual survival for midpoints of all life stages.
}
\examples{
pop <- cbind(0:5,
             sort(runif(6,0,10), decreasing=TRUE),
             sort(c(0, 0, 0, runif(3,0,12))))
poppar(pop)
poppar(pop, 50)

}
\references{
Elzinga, C. Laboratory Manual for Honors Organismal Biology, Michican State University. Retrieved from \url{https://msu.edu/course/lbs/158h/manual.html}, February 2019.

Krebs, C. J. (2009). Ecology. San Francisco, Pearson.
}
\seealso{
\href{https://cran.r-project.org/package=MortalityTables}{MortalityTables}

\href{https://cran.r-project.org/package=LifeTables}{LifeTables}
}
\author{
Ölvir Styrmisson
}
