% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poppar.R
\name{poppar}
\alias{poppar}
\title{Population parameters from lifetable}
\usage{
poppar(pop, K)
}
\arguments{
\item{pop}{A 3 column table containing age, abundance at each life stage, whether relative or absolute, and offspring produced at lifestage.}

\item{K}{Estimated carrying capacity of the envirorment where the represented population lives.}
}
\value{
R0,Tc,r,birth-ex: Current population growth, average generation time, intrinsic population growth and life expectancy at birth (if age 0 is at the top of a given table and life stages are seperated by a constant interval, e.g. 1 year) all expressed in the given time unit.
}
\description{
This function gives various point estimates of current population growth.
}
\details{
K, carrying capacity, is optional. If it is not provided it is assumed the observed growth is not contrained by the population density.
}
\note{
Calculation of estimated parameters:
R0 = Σ(lx x bx)
Gc = Σ(lx x bx x age) / R0
r(1) =  ln(R0) / Gc
r(2) = ((dN/dt) x K) / (N x (K - N))
ex = Tx / lx

Where R0 is the basic reproduction rate, lx proportion of living individuals at age x compared to age 0, bx numbers of offspring produced by individual at age x, Gc the average generation length, age the age in a given time unit, r the instantaneous multiplication rate, defined without or with K, Tx is a sum of proportion of individuals surviving at the mindpoint of present life stage (here age 0) and future life stages and ex is the execpted life length at age 0.

Some other needed values:
lx = Ni/N0
bx = Noi/Ni
Lx = (lxi + lxi+1)/2
Tx = Σ Lx

Ni is the number of individuals at a particular live stage, Noi is the offspring produced at a particular live stage and Lx is the proportion of individual survivial for midpoints of all livestages.
}
\examples{
pop = (cbind((0:5),sort(runif(6,0,10),decreasing=TRUE),sort(c(0,0,0,runif(3,0,12)))))
poppar

}
\references{
Elzinga, C. Laboratory Manual for Honors Organismal Biology, Michican State University. Retrieved from https://msu.edu/course/lbs/158h/manual.html, February 2019
Krebs, C. J. (2009). Ecology. San Francisco, Pearson.
}
\seealso{
\href{https://cran.r-project.org/package=MortalityTables/}{MortailityTables}
 \href{https://cran.r-project.org/package=LifeTables/}{LifeTables}
}
\author{
Ölvir Styrmisson
}
