% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r2.R
\name{r2}
\alias{r2}
\title{Calculate R2}
\usage{
r2(x, y)
}
\arguments{
\item{x}{a numeric vector.}

\item{y}{a numeric vector.}
}
\value{
R2 as a number.
}
\description{
This function calculates R2, the coefficient of determination, from a linear
regression.
}
\examples{
r2(cars$speed, cars$dist)

cor(cars$speed, cars$dist)^2

summary(lm(dist~speed, data=cars))$r.squared

}
\seealso{
\code{\link{lm}}
}
\author{
Hafrún Gunnarsdóttir
}
