% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twice.R
\name{twice}
\alias{twice}
\title{Multiply by Two}
\usage{
twice(x)
}
\arguments{
\item{x}{a number.}
}
\value{
A number, two times the value of x.
}
\description{
This function multiplies x by 2.
}
\examples{
twice(10)

twice(pi)

}
\seealso{
\link{Arithmetic}
}
\author{
Arni Magnusson
}
